fix: compatibility with pypy
Pypy isn't smart enough to look for the last single quote in the line and that would be where we'd intend the string to terminate at.
Now you can use Pypy, not sure if upstream would let you because I haven't tested, but it doesn't seem faster than CPython, I'm using Pypy `3.10` version `7.3.17`.