fix: Remove comments from vscode debug config file
These make GitHub scream because I didn't override the file type in `.gitattributes` to make it recognize it as JSON with Comments / JSONC instead of JSON. Too lazy to add the `.gitattributes` file, so this works.
Also probably shouldn't be there anyways because it's not meant for anything other than the template.